[check_url.pyRefactoring code to support packages
