+++ .vscode/settings.json
@@ -1,0 +1,8 @@
+{
+    "python.testing.nosetestsEnabled": false,
+    "python.testing.pytestEnabled": true,
+    "python.testing.cwd": "",
+    "python.testing.pytestArgs": [
+        "-v"
+    ]
+}
