+++ test/test_check_url_check_output.py
@@ -1,45 +1,43 @@
-import pytest
-import logging
-from check_url import check_url_output
+#!/usr/bin/python
 
+import sys 
+
+import pytest
+
+import logging
+from check_xpath import check_xpath_class
 
 URL_LOCAL_SUOTA_OK = """path : Device/Services/AdvancedFwUpdate/URL
 value : 'http://pi-fry.home/TEST2/WHW6'
 """
 
-def test_check_output_1():
-    assert check_url_output(URL_LOCAL_SUOTA_OK) == True
-
 URL_LOCAL_SUOTA_KO = """path : Device/Services/AdvancedFwUpdate/URL
-value : 'https://wifi-suota-test.cpe-mgt.bt.com/TEST2/WHW6'
+value : 'http://pi-fry.home/TEST2/WHW7'
 """
-def test_check_output_2():
-    assert check_url_output(URL_LOCAL_SUOTA_KO) == False
 
-URL_LOCAL_SUOTA_KO_1 = """path : Device/Services/AdvancedFwUpdate/URL
-value : 'https://wifi-suota-test.cpe-mgt.bt.comTEST2/WHW6'
-"""
-def test_check_output_3():
-    assert check_url_output(URL_LOCAL_SUOTA_KO_1) == False
+XPATH = "Device/Services/AdvancedFwUpdate/URL"
+EXPECTED_VALUE = "http://pi-fry.home/TEST2/WHW6"
 
-URL_LOCAL_SUOTA_KO_2 = """path : Device/Services/AdvancedFwUpdate/URL
-vlue : 'https://wifi-suota-test.cpe-mgt.bt.comTEST2/WHW6'
-"""
-def test_check_output_4():
-    assert check_url_output(URL_LOCAL_SUOTA_KO_2) == False
+disc = {"ip":"192.168.10.11", "role": "MASTER", "name":"STUDY", "username":"root", "password":"root"}
 
+# @pytest.fixture
+# def mock_runCommand(mocker):
+#     return mocker.
 
-URL_LOCAL_SUOTA_KO_3 = """path : Device/Services/AdvancedFwUpdate/URL
-"""
-def test_check_output_5():
-    assert check_url_output(URL_LOCAL_SUOTA_KO_3) == False
+def test_check_output_1(mocker):
+    mocker.patch('tools.tools.runCommand', return_value = (URL_LOCAL_SUOTA_OK.encode('utf8'),True))
+    
+    xpath_check = check_xpath_class(XPATH, EXPECTED_VALUE)
+    assert xpath_check.check_xpath(disc) == True
 
-URL_LOCAL_SUOTA_KO_4 = """value : 'https://wifi-suota-test.cpe-mgt.bt.comTEST2/WHW6'
-"""
-def test_check_output_6():
-    assert check_url_output(URL_LOCAL_SUOTA_KO_4) == False
+def test_check_output_2(mocker):
+    mocker.patch('tools.tools.runCommand', return_value = (URL_LOCAL_SUOTA_OK.encode('utf8'),False))
 
-URL_LOCAL_SUOTA_KO_5 = """vlue : 'https://wifi-suota-test.cpe-mgt.bt.comTEST2/WHW6'
-"""
-def test_check_output_7():
-    assert check_url_output(URL_LOCAL_SUOTA_KO_5) == False
+    xpath_check = check_xpath_class(XPATH, EXPECTED_VALUE)
+    assert xpath_check.check_xpath(disc) == False
+
+def test_check_output_3(mocker):
+    mocker.patch('tools.tools.runCommand', return_value = (URL_LOCAL_SUOTA_KO.encode('utf8'),True))
+
+    xpath_check = check_xpath_class(XPATH, EXPECTED_VALUE)
+    assert xpath_check.check_xpath(disc) == False
